feature/ASAA-18 clean up places
