+++ data/build.gradle.kts
@@ -132,8 +132,6 @@ dependencies {
         devConfigurations = devConfigurations,
         productionConfiguration = productionReleaseImplementation
     )
-    // Places
-    placesApiDependencies()
 
     // Test
     junitDependencies()
